refactor(#93): 북마크 관련 내용이 저장된 패키지이름을 (favorite -> bookmark)로 변경
