+++ src/main/java/com/sideproject/hororok/auth/application/AuthService.java
@@ -8,8 +8,8 @@ import com.sideproject.hororok.auth.dto.OAuthMember;
 import com.sideproject.hororok.auth.dto.request.TokenRenewalRequest;
 import com.sideproject.hororok.auth.dto.response.AccessAndRefreshTokenResponse;
 import com.sideproject.hororok.auth.dto.response.AccessTokenResponse;
-import com.sideproject.hororok.favorite.domain.BookmarkFolder;
-import com.sideproject.hororok.favorite.domain.BookmarkFolderRepository;
+import com.sideproject.hororok.bookmark.domain.BookmarkFolder;
+import com.sideproject.hororok.bookmark.domain.BookmarkFolderRepository;
 import com.sideproject.hororok.member.domain.Member;
 import com.sideproject.hororok.member.domain.MemberRepository;
 import lombok.RequiredArgsConstructor;
