+++ src/main/java/com/sideproject/hororok/bookmark/domain/BookmarkFolder.java
@@ -1,0 +1,60 @@
+package com.sideproject.hororok.bookmark.domain;
+
+import com.sideproject.hororok.bookmark.dto.request.BookmarkFolderUpdateRequest;
+import com.sideproject.hororok.member.domain.Member;
+import jakarta.persistence.*;
+import lombok.Getter;
+
+import static jakarta.persistence.FetchType.*;
+import static jakarta.persistence.GenerationType.*;
+
+
+@Getter
+@Entity
+@Table(name = "bookmark_folders")
+public class BookmarkFolder {
+
+    @Id
+    @GeneratedValue(strategy = IDENTITY)
+    private Long id;
+
+    @Column(name = "name", nullable = false)
+    private String name;
+
+    @Column(name = "color", nullable = false)
+    private String color;
+
+    @Column(name = "is_visible", nullable = false)
+    private Boolean isVisible;
+
+    @Column(name = "is_default_folder", nullable = false)
+    private Boolean isDefaultFolder;
+
+    @ManyToOne(fetch = LAZY)
+    @JoinColumn(name = "members_id")
+    private Member member;
+
+    protected BookmarkFolder() {
+    }
+
+    public BookmarkFolder(final String name, final String color,
+                          final Boolean isVisible,
+                          final Boolean isDefaultFolder,
+                          final Member member) {
+        this.name = name;
+        this.color = color;
+        this.isVisible = isVisible;
+        this.isDefaultFolder = isDefaultFolder;
+        this.member = member;
+    }
+
+    public void change(BookmarkFolderUpdateRequest request) {
+        this.name = request.getName();
+        this.color = request.getColor();
+        this.isVisible = request.getIsVisible();
+    }
+
+    public void changeVisible() {
+        this.isVisible = !isVisible;
+    }
+}
