+++ src/main/java/com/sideproject/hororok/bookmark/domain/BookmarkFolderRepository.java
@@ -1,0 +1,12 @@
+package com.sideproject.hororok.bookmark.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface BookmarkFolderRepository extends JpaRepository<BookmarkFolder, Long> {
+
+    Long countByMemberId(Long memberId);
+    List<BookmarkFolder> findByMemberId(Long memberId);
+
+}
