+++ src/main/java/com/sideproject/hororok/bookmark/dto/request/BookmarkFolderUpdateRequest.java
@@ -1,0 +1,24 @@
+package com.sideproject.hororok.bookmark.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class BookmarkFolderUpdateRequest {
+
+    private Long folderId;
+    private String name;
+    private String color;
+    private Boolean isVisible;
+
+    private BookmarkFolderUpdateRequest() {
+    }
+
+    public BookmarkFolderUpdateRequest(
+            final Long folderId, final String name,
+            final String color, final Boolean isVisible) {
+        this.folderId = folderId;
+        this.name = name;
+        this.color = color;
+        this.isVisible = isVisible;
+    }
+}
