+++ src/main/java/com/sideproject/hororok/bookmark/dto/request/BookmarkSaveRequest.java
@@ -1,0 +1,18 @@
+package com.sideproject.hororok.bookmark.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class BookmarkSaveRequest {
+
+    private Long cafeId;
+    private Long folderId;
+
+    private BookmarkSaveRequest() {
+    }
+
+    public BookmarkSaveRequest(final Long cafeId, final Long folderId) {
+        this.cafeId = cafeId;
+        this.folderId = folderId;
+    }
+}
