+++ src/main/java/com/sideproject/hororok/bookmark/exception/DefaultFolderDeletionNotAllowedException.java
@@ -1,0 +1,12 @@
+package com.sideproject.hororok.bookmark.exception;
+
+public class DefaultFolderDeletionNotAllowedException extends RuntimeException{
+
+    public DefaultFolderDeletionNotAllowedException(final String message) {
+        super(message);
+    }
+
+    public DefaultFolderDeletionNotAllowedException() {
+        this("기본 폴더는 삭제할 수 없습니다.");
+    }
+}
