+++ src/main/java/com/sideproject/hororok/bookmark/exception/NoSuchBookmarkException.java
@@ -1,0 +1,12 @@
+package com.sideproject.hororok.bookmark.exception;
+
+public class NoSuchBookmarkException extends RuntimeException{
+
+    public NoSuchBookmarkException(final String message) {
+        super(message);
+    }
+
+    public NoSuchBookmarkException() {
+        this("존재하지 않는 북마크 입니다.");
+    }
+}
