+++ src/main/java/com/sideproject/hororok/bookmark/exception/NoSuchFolderException.java
@@ -1,0 +1,12 @@
+package com.sideproject.hororok.bookmark.exception;
+
+public class NoSuchFolderException extends RuntimeException{
+
+    public NoSuchFolderException(final String message) {
+        super(message);
+    }
+
+    public NoSuchFolderException() {
+        this("존재하지 않는 폴더입니다.");
+    }
+}
