--- src/main/java/com/sideproject/hororok/favorite/exception/NoSuchFolderException.java
@@ -1,12 +1,0 @@
-package com.sideproject.hororok.favorite.exception;
-
-public class NoSuchFolderException extends RuntimeException{
-
-    public NoSuchFolderException(final String message) {
-        super(message);
-    }
-
-    public NoSuchFolderException() {
-        this("존재하지 않는 폴더입니다.");
-    }
-}
