+++ src/test/java/com/sideproject/hororok/common/annotation/ControllerTest.java
@@ -5,10 +5,10 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.sideproject.hororok.auth.application.AuthService;
 import com.sideproject.hororok.auth.application.OAuthClient;
 import com.sideproject.hororok.auth.presentation.AuthController;
-import com.sideproject.hororok.favorite.application.BookmarkFolderService;
-import com.sideproject.hororok.favorite.application.BookmarkService;
-import com.sideproject.hororok.favorite.presentation.BookmarkController;
-import com.sideproject.hororok.favorite.presentation.BookmarkFolderController;
+import com.sideproject.hororok.bookmark.application.BookmarkFolderService;
+import com.sideproject.hororok.bookmark.application.BookmarkService;
+import com.sideproject.hororok.bookmark.presentation.BookmarkController;
+import com.sideproject.hororok.bookmark.presentation.BookmarkFolderController;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
